+++ assets/styles/Colors.js
@@ -1,0 +1,20 @@
+export default {
+  secondary:'#41FFE1',
+  primary:'#363F46',
+  grey:'#728694',
+  tintColor:'#2f95dc',
+};
+export const notifications = {
+  errorBackground: 'red',
+  errorText: '#fff',
+  warningBackground: '#EAEB5E',
+  warningText: '#666804',
+  noticeBackground: '#2f95dc',
+  noticeText: '#fff',
+}
+
+export const TabBar = {
+  tabIconDefault: '#ccc',
+  tabIconSelected: '#2f95dc',
+  tabBar: '#fefefe',
+}
