+++ assets/styles/Spacing.js
@@ -1,0 +1,5 @@
+export default {
+    containerSpacing: {
+        padding: 20,
+    }
+}
