+++ assets/styles/index.js
@@ -1,0 +1,4 @@
+import * as Colors from './Colors'
+import * as Spacing from './Spacing'
+
+export { Colors, Spacing }
