--- components/StyledText.js
@@ -1,9 +1,0 @@
-import React from 'react';
-import {Text} from 'react-native';
-
-export function MonoText() {
-    return (
-        <Text style={[props.style, {fontFamily: 'space-mono'}]}>yoooo</Text>
-            )
-          }
-
