+++ screens/LoginScreen.js
@@ -6,6 +6,7 @@ import {CallApi} from '../components/callApi';
 import t from "tcomb-form-native";
 import LoginUser from "../constants/User"
 import {FacebookButton} from "../components/FacebookButton"
+import {Colors, Spacing} from "../assets/styles";
 
 const Form = t.form.Form
 
@@ -30,10 +31,10 @@ export default class LoginScreen extends Component {
 }
 const styles = StyleSheet.create({
     container: {
+        flex: 1,
+        ...Spacing.default.containerSpacing,
         justifyContent: 'center',
-        marginTop: 50,
-        padding: 20,
-        backgroundColor: '#ffffff',
+        backgroundColor:Colors.default.primary
     },
 });
